requirements.txt
numpy
dateparser
Pillow
RPi.GPIO 
spidev